refactor: use env-based supabase client
