+++ src/lib/supabase/promptBuilder.ts
@@ -1,4 +1,4 @@
-import { supabase } from '@/lib/supabase';
+import { supabase } from '@/lib/supabaseClient';
 
 export interface DialogueResponse {
   initial_dialogue_templates: {
